Change "Background" to "Highlight"
The word "Highlight" is better suited to english users. 

"Background" is the term used in HTML; however, a common English user would think that the background color would change the background of the entire editor instead of the chosen text. The word "Highlight" more accurately describes the effect of changing the background behind a line of text.